Add admin login form and integrate into blog
