+++ pages/blog.js
@@ -4,11 +4,14 @@ import ArticleCard from '../components/ArticleCard';
 import Navbar from '../components/navbar';
 import Footer from '../components/Footer';
 import SponsorsBar from '../components/Sponsors';
+import AdminLoginForm from '../components/AdminLoginForm';
 import Head from 'next/head';
+import Link from 'next/link';
 
 export default function Blog() {
     const { articles } = useArticles();
     const [posts, setPosts] = useState([]);
+    const [isAdmin, setIsAdmin] = useState(false);
 
     useEffect(() => {
         async function load() {
@@ -21,6 +24,7 @@ export default function Blog() {
             }
         }
         load();
+        setIsAdmin(document.cookie.includes('admin-auth=true'));
     }, [articles]);
 
     if (!posts || posts.length === 0) {
@@ -40,6 +44,16 @@ export default function Blog() {
                     <ArticleCard key={article.id} article={article} isLarge={true}/>
                 ))}
             </div>
+            {!isAdmin ? (
+                <div className="my-8">
+                    <h2 className="text-xl font-bold mb-2">Admin Login</h2>
+                    <AdminLoginForm />
+                </div>
+            ) : (
+                <div className="my-8 text-center">
+                    <Link href="/admin" className="bg-blue-500 text-white px-4 py-2 rounded">Go to Admin Dashboard</Link>
+                </div>
+            )}
             <SponsorsBar />
             <Footer />
         </div>
